editor 필터, 약력 추가, pagination 추가

+++ highlight/portfolio/views.py
@@ -2,42 +2,55 @@ from django.shortcuts import render, get_object_or_404, redirect
 from .models import Portfolio, Review
 from django.utils import timezone
 from .forms import PortfolioForm
+from django.core.paginator import Paginator
 
 # Create your views here.
-def list(request):
-    portfolio = Portfolio.objects
-    return render(request, 'portfolio/portfolio_list.html', {'portfolio':portfolio})
+def main(request):
+    portfolio = Portfolio.objects.all()
+    paginator = Paginator(portfolio, 5)
+    page = request.GET.get('page')
+    portfolio = paginator.get_page(page)
+    return render(request, 'portfolio/portfolio_main.html', {'portfolio':portfolio})
 
 def detail(request, portfolio_id):
     portfolio_detail = get_object_or_404(Portfolio, pk = portfolio_id)
     return render(request, 'portfolio/portfolio_detail.html', {'portfolio' : portfolio_detail})
 
 def portfolio_new(request):
-    if request.method == "POST":
-        form = PortfolioForm(request.POST)
-        if form.is_valid():
-            portfolio = form.save(commit=False)
-            # request.user.is_authenticated
-            portfolio.editor = request.user
-            portfolio.save()
-            return redirect('detail', portfolio_id = portfolio.pk)
-    else:
-        form = PortfolioForm()
-    return render(request, 'portfolio/portfolio_new.html', {'form': form})
+    if request.user.user_type == 'editor': 
+        if request.method == "POST":
+            form = PortfolioForm(request.POST)
+            if form.is_valid():
+                portfolio = form.save(commit=False)
+                # request.user.is_authenticated
+                portfolio.editor = request.user
+                portfolio.save()
+                return redirect('detail', portfolio_id = portfolio.pk)
+        else:
+            form = PortfolioForm()
+        return render(request, 'portfolio/portfolio_new.html', {'form': form})
+    else :
+        return redirect('main')
 
 def portfolio_edit(request, portfolio_id):
     portfolio = get_object_or_404(Portfolio, pk = portfolio_id)
-    if request.method == "POST":
-        form = PortfolioForm(request.POST, instance = portfolio)
-        if form.is_valid():
-            portfolio = form.save(commit=False)
-            portfolio.save()
-            return redirect('detail', portfolio_id = portfolio.pk)
+    if request.user.user_type == 'editor':
+        if request.method == "POST":
+            form = PortfolioForm(request.POST, instance = portfolio)
+            if form.is_valid():
+                portfolio = form.save(commit=False)
+                portfolio.save()
+                return redirect('detail', portfolio_id = portfolio.pk)
+        else:
+            form = PortfolioForm(instance = portfolio)
+        return render(request, 'portfolio/portfolio_edit.html', {'form':form})
     else:
-        form = PortfolioForm(instance = portfolio)
-    return render(request, 'portfolio/portfolio_edit.html', {'form':form})
+        return redirect('detail', portfolio_id = portfolio.pk)
 
 def portfolio_delete(request, portfolio_id):
     portfolio = get_object_or_404(Portfolio, pk=portfolio_id)
-    portfolio.delete()
-    return redirect('list')
+    if request.user.user_type == 'editor':
+        portfolio.delete()
+        return redirect('main')
+    else:
+        return redirect('detail', portfolio_id = portfolio.pk)
